added useFormWatcher
